Added rmgpy.pdep.cse module for the chemically-significant eigenvalues method.
This is the most accurate of the k(T,P) estimation methods, and is
"exact" as long as there are enough chemically distinct eigenmodes.
However, this condition is unlikely to hold for networks of even modest
size as you move away from the high-pressure limit, especially at high
temperature. This is generally treated as representing a need to lump
k(T,P) values together, but it's not clear how to do this in an
automatic manner.