Make MoleculeDrawer.render() into a public method.
This way we can call it from outside the class. We will use this in our
updated network drawing algorithm.

Also changed the location of the white background drawing to be outside
of the render() method. Other things that call this method might not want
(or might already have) a white background.