Added NetworkDrawer class to rmgpy.pdep.draw module.
This class provides an updated implementation of the network drawing
functionality. Like the recently refactored MoleculeDrawer and
ReactionDrawer classes, it has been changed to its own class to allow
the user to more easily give options to the drawing algorithm.

The network drawing algorithm itself is improved in several ways. First,
wells with extra wide labels are given additional space so that the
labels do not overlap horizontally, as they might for large structures or
long names. Second, the true bounding rect of each well is computed and
used to more intelligently squish wells horizontally so they don't
overlap vertically. Finally, the old algorithm was drawing the labels as
raster images instead of vector objects; the new version avoids this by
drawing everything on to the final surface.