Added CanTherm class to rmgpy.cantherm.main module.
The CanTherm class represents an instance of a CanTherm job, used to
compute properties of species and reactions. Wrapping this functionality
in a class enables us to more easily generate jobs programmatically as
well as using input files. Methods for parsing the command-line arguments,
initializing the logging system, and running the job are provided.