Added ThermoJob class to rmgpy.cantherm.thermo module.
A ThermoJob is used to generate a thermodynamics model for a species
by evaluting its statistical mechanical degrees of freedom. Either a
Wilhoit or NASA model can be fit; as usual, the NASA is fit from the
Wilhoit fit. A plot comparing the fitted thermo model to the thermo
values evaluated from the statistical mechanical modes is also generated
if matplotlib is installed.