Added PressureDependenceJob class to rmgpy.cantherm.pdep module.
The PressureDependenceJob class provides functionality for computing the
pressure-dependent rate coefficients k(T,P) for a unimolecular reaction
network and fitting them to a kinetics model. Plots of the computed
k(T,P) values and the fitted kinetics models are produced if matplotlib
is installed. A depiction of the network is produced if Cairo and its
Python wrapper are installed.

This is largely a merger of MEASURE into CanTherm.