Updated loadInputFile() in rmgpy.cantherm.input module.
The updated version of the loadInputFile() function loads the CanTherm
input files for the statistical mechanics, thermodynamics, kinetics, and
pressure dependence jobs. Some functionality that was duplicated in other
files (e.g. the projecting out of the hindered rotors from the force
constant matrix) has been removed from this module.