Added prettify() function to rmgpy.cantherm.output module.
This function takes repr() output for a Python object and "prettifies" it
by spreading it onto multiple lines with appropriate indentation. This
function is used in each job to generate the output file. The
implementation involves walking the Python abstract syntax tree for a
given snippet of code, which is easily done using the ast module.

This may be useful outside of CanTherm as well, but as CanTherm is the
only thing using it at the moment, I am placing it inside the cantherm
subpackage for now.