Added rmgmode attribute to PressureDependenceJob class.
"RMG mode" turns on a number of additional approximations that are
generally not needed for detailed pdep calculations but are needed in
RMG due to the incomplete information available. Among these:

* Densities of states are not computed for product channels; this was a
  big time penalty in RMG jobs due to the large number of product
  channels generated by RMG

* Arbitrary rigid rotor moments of inertia are included in the active
  modes; these cancel in the ILT and equilibrium expressions

* k(E) for each path reaction is computed in the direction A -> products,
  where A is always an explored isomer; the high-P kinetics are reversed
  if necessary for this purpose

* Thermodynamic parameters are always used to compute the reverse k(E)
  from the forward k(E) for each path reaction

The idea is to make the default case the one that does not use these
approximations, and only turn them on when explicitly requested by the
user. Note that the rmgmode flag can be set in the CanTherm input file,
so that we can debug pdep calculations from within RMG jobs.