Improvements to the inverse Laplace transform method.
The new version lumps the T^n piece into the A factor for n < 0.2. This
is because I found that the algorithm for the T^n piece does not nicely
approach that of the n = 0 case as n approaches zero. In fact, this
change actually has less error in the computed k(E) than trying to
compute it "correctly". I suspect the algorithm has poor numerical
properties for small n, especially since we use very large E/R values.

The new version also better handles the case when a semilogarithmic
interpolation is requested between a zero and a nonzero value. In this
case an extrapolation using the slope of the neighboring grains is
used instead. The result is a somewhat smaller discretization error in
the computed k(E).