Delete intermediate arrays used during pdep calculations in RMG jobs.
These arrays can use a fair amount of memory, so we can't really afford
to keep them around in RAM during long RMG jobs. In the future we might
want to look into caching this information to disk rather than just
discarding it in order to avoid having to recompute it, but that's a
task for another time.