Don't copy PressureDependenceJob when updating a network.
There is no need to make a copy of the job settings, as we can just
swap in the network instead. The copy as previously implemented turned
out to be rather expensive (~10% of total job), so this is potentially
valuable savings.

This may have to change again when we parallelize.