Calculate densities of states by direct count in RMG jobs.
The inverse Laplace transform of the partition function by the method of
steepest descents is probably better for detailed studies of individual
networks, especially for proper treatment of the hindered rotors.
However, in RMG jobs we are making enough approximations that the
additional CPU cost of this method is not appropriate. The direct count
method as currently implemented is significantly faster to evaluate,
as it does not involve a series of optimizations.

Note that, to use direct count, we must treat the hindered rotors purely
classically. Since the hindered rotors are approximated anyway, this is
not much of a loss.