Added commonUnits attribute to UnitType class for faster checking.
Basically commonUnits is a list of the most common units of a given unit
type, units that are known to be handled properly by the quantities
package. This is used as a cache to avoid calling the slow quantities
package where possible.

I've implemented common units for the thermo parameters (energy,
enthalpy, free energy, heat capacity, and entropy). and the effect is
quite dramatic. I've seen a full 2x speed boost from just this commit
alone.