Use a constant value of B when fitting Wilhoit polynomial in RMG jobs.
This more closely mimics the behavior of RMG-Java, and is a bit faster
than allowing B to optimize. As far as I can tell, the additional
fitting error is small, especially given the approximate nature of most
of the Cp estimates anyway.

A fixed value of B = 1000 K was (somewhat arbitrarily) chosen.