Fix KineticsFamily.__getAverageKinetics() to work with many expressions.
The previous implementation ran the risk of overflowing the value of
the A factor when many Arrhenius expressions were averaged together.
The new implementation works with log A instead, which, while slower,
is also more robust.

(This was only noticed recently when importing the much larger, sulfur-
containing RMG database from RMG-Java.)