Added maximumAtoms option for pressure dependence in RMG jobs.
This feature allows you to only run pressure dependence for reactions
for only the small-molecule chemistry in your model (which is more
likely to be pressure-dependent than the heavy molecules). You can use
the maximumAtoms option in the pressureDependence() block of the input
file to control the cutoff between small-molecule and large-molecule
chemistry; note that this is in total atoms, not heavy atoms. This
should help alleviate speed issues when building models for very
heavy initial species (e.g. diesel surrogates, JP-10, etc.)