Only use defined ranges of density and sum of states for interpolation mapping.
When computing k(E) using RRKM theory, we manually convolve the tunneling
function kappa(E) into the transition state density of states to obtain
the transition state sum of states. This sum of states is already shifted
to the correct transition state energy, which means it will have some
zero values. This fix eliminates a log(0) error which was occurring while
setting up the semilogarithmic interpolator.