Always keep the exergonic pdep reaction from the activated species algorithm.
Since RMG-Py uses the activated species algorithm of Matheu, it generates
A + B -> C and C -> A + B from separate partial networks. Before I was
trying to keep the rate from the "more explored" network, but now I think
that is too ill-defined to be useful. Instead, I am now simply keeping the
exergonic (dGrxn < 0) direction; this is consistent with our handling of
reaction families that are their own reverse.

I've also updated the check for thermo consistence between the forward and
reverse net reactions to print the discrepancy of the deviation so the
user at least knows about it.