Always check for duplicate reactions before saving Chemkin file.
This helps ensure that the generated Chemkin file is valid. If the duplicate
status of a reaction must be set, a warning is printed to the log. In the
future we should probably have these reactions already marked as duplicate
before we reach this point, so that this warning message is not printed.