Delete unused and broken 'getData(self, label)' method of Database.
As far as I can tell, this method was introduced in c5ee064c17a8bfe9164a694508825a30195b302f
with the bug (line 745 would give NameError: global name 'key' is not defined)
but has never been used in this project. It is currently not used in the database
or the website.
I can't tell from the docstring quite what it was intended to do, so I'm just deleting it.
If we need it back, revert this commit.