Updated the SimpleReactor unit test to be functional again.
Besides fixing some imports and some units, I've added a few assertions so
that the unit test actually checks something. I've also commented out the
plotting code so that we can run this test as part of the entire batch
without needing interactivity from the user (to close the plot window).