Removed remaining modules from unittest folder.
Most of these tests have either been replaced by updated versions or are so
old that they would need to be almost completely rewritten.

All new unit tests should be placed in modules in the rmgpy folder such
that we can run all of them with the command

$ find rmgpy -name *Test.py -exec nosetests2 {} +

At the moment we have 655 passing tests out of 664 (98.6%). All of the
failing tests involve calculating molecule symmetry numbers and other
properties from the 2D graph representation, and show the limitations of
this approach.