Better handling of species labels when loading Chemkin file.
In RMG-generated input files, each species in the list of species is given
its own line, so that a comment with the "full" name of the species may be
provided if we have need to truncate the name to fit in the space
limitations. When loading one of these input files, we want to recover
the full name; this commit adds functionality for doing so.

This commit also adds functionality for extracting the species index from
the name. Again, this only applies to RMG-generated Chemkin files, though
hopefully I've designed it to work for generic Chemkin files as well.