Changed format of key for storing entries in some kinetics databases.
Previously the index of each entry was being used as the key. In principle
this should work fine, as each entry should have a unique index. In
practice, however, there are currently many entries in the database
(especially for H_Abstraction) that share an index, mainly because of
the sulfur rules. This commit enables the ability to handle multiple
entries with the same index (as long as they have different labels).