Added first version of tool for merging RMG-generated models.
Like many of the existing web tools, you need the Chemkin file and the
species dictionary for each model you are merging. The models to be
merged should be given in order of priority, since, when common species
and reactions are encountered, the earlier model is given priority.
You can also merge more than two models at once if you are so inclined.

The standalone script mergeModels.py could use a few more options (e.g.
the ability to specify where the output files are written), but it is
usable. The actual merging is done in the ReactionModel class, so it
should be easy to set this up as a web tool as well.