Generate reactions now allows constraints on the generated species.
These are passed as (optional) keyword options to various methods used to
generate reactions. The keywords allow for control of the maximum number
of C, H, O, N, Si, and S atoms, the total number of heavy (non-H) atoms,
and the total number of radicals.

Note that RMG-Java also allows for setting the maximum number of cycles,
but I've left this out for now since finding cycles is considerably more
expensive than counting elements and radicals.