Added generatedSpeciesConstraints() block to RMG input file.
This optional block allows the user to set constraints on the species
generated during RMG jobs. The allowed constraints are:
maximumCarbonAtoms, maximumHydrogenAtoms, maximumOxygenAtoms,
maximumNitrogenAtoms, maximumSiliconAtoms, maximumSulfurAtoms,
maximumHeavyAtoms, and maximumRadicalElectrons.

These constraints are stored in the new reactionGenerationOptions
attribute of the CoreEdgeReactionModel class. The contents of this dict
are passed to KineticsDatabase.generateReactions() as the options so
that they may be enforced during reaction generation.

The expectation is that, if used, this should provide a significant speed
boost to many RMG jobs by eliminating graph isomorphism evaluations for
large molecules, which are known to be slow.