Converted rmgpy.data.kinetics into subpackage.
The module rmgpy.data.kinetics was nearing 4000 lines of code, which was
by far the largest individual module in the project. There's enough
different functionality to warrant splitting into several modules.
Grouping these under a new rmgpy.data.kinetics subpackage enables other
modules to continue interacting as if it was still a single file.

The kinetics depository, library, and family functionality has been
split into separate modules inside the rmgpy.data.kinetics subpackage.
A module for common functionality is also provided. The main
KineticsDatabase class is placed in the __init__.py since it depends on
all the other modules in the subpackage.