Better monocyclic ring corrections in group additivity thermo estimation.
The previous implementation oddly made no attempt to account for ligands
when determining the ring correction to apply. The new version finds the
ring correction at each atom in the ring, and applies the one with the
largest H298, as this indicates larger ring strain. The new algorithm
works perfectly for fulvene.