MultiPDepArrhenius requires identical pressures in all PDepArrhenius objects.
This is so we can evaluate the k(T,P) values by evaluating the sum of the
k(T,P_1) and k(T,P_2) values across the different PDepArrhenius
expressions, then interpolating logarithmically in pressure. The idea is
that you need this sum-of-PLOG format to accurately capture the very
non-Arrhenius behavior of some pressure-dependent kinetics.