Entries in KineticsRules now grouped by template.
For a given template, you no longer need to iterate over all of the
entries to find the possible duplicates for that template. Instead, you
can get the entire list directly using the template as the key to the
entries dict. To get the entire list of values as before, use the new
KineticsRules.getEntries() method.