KineticsRules.getRule() now returns None if no rule is found.
This is for speed. When we are filling in the kinetics rate rules by
averaging up, we use getRule() a lot to check if rules exist. When they
didn't exist before, a ValueError was raised, which was quite expensive
because it occurred so frequently. This change causes a significant
speedup in the filling in of the kinetics rate rules.

Note that KineticsFamily.getRateRule() still raises ValueError when
KineticsRules.getRule() returns None.