Added checking of common units when parsing rate coefficients.
We've used this trick with other unit types to get a speed boost, but we
had forgotten to include it with the special case of the rate coefficient.
Since this is an extremely common unit type to parse in RMG, I expect this
change to come with a noticeable speed boost for a few areas.

Note that I've only added unimolecular and bimolecular rate coefficient
units, as other molecularities are too rare to be worth it.