In reservoir state method, use the deepest well to choose the matrix bandwidth.
We use a banded matrix in the reservoir state method to accelerate the
linear solve. The halfbandwidth of this matrix is chosen by neglecting
P(E,E') values below a certain threshold. It is important to choose the
deepest isomer well for this, as it gives the largest halfbandwidth and
therefore the safest banded matrix.

I noticed this issue in a PES where I had a very shallow adduct as the
first isomer well. Since the previous implementation used the first isomer
only to choose the halfbandwith, it was choosing a value that was far too
small for the other isomer well.