Implemented preliminary version of CSE lumping algorithm.
The lumping algorithm applies when you are deficient by at least one
reaction timescale due to blending into the collision timescales. It
eliminates one or more isomer wells from consideration and returns the
subset of the k(T,P) values for the remaining configurations. This might
be useful for investigations of individual networks in detail.

The priority of lumping is indicated with the lumpingOrder parameter;
probably you want to lump the shallowest isomer wells first most of the
time. (Think pseudo-steady state approximation, not rapid equilibrium
approximation.) If the lumping order is not specified, or if the list is
too short to account for all of the missing eigenmodes, the CSE method
fails as before.

I don't know if there is a good way to "unlump" to fill in the missing
k(T,P) values.