Some refactoring of the k(E) consistency check.
In the new implementation, kf0 and kr0 should give the same value in both
the RRKM and ILT modes. This way, the equilibrium constant is given by
Qprod / Qreac (formerly Qr / Qf) as expected.

The previous check was working fine; this is just a bit clearer as to
what is actually going on.