Deprecated use of frequency scale factor in CanTherm species files.
Previously you could specify the frequency scale factor in either or both
of the job input file or the species files, which is confusing and
redundant. I decided to keep the one in the job input file, since it
probably ought to be the same for all species in your job. A warning is
printed if this value is missing from the job input file or found in a
species file, to encourage users to make the change.