Added ScanLog class for loading/saving scan energies from a text file.
This provides the option of supplying the scan energies manually in a
text file instead of reading them from a Gaussian log file. The file is
also automatically saved by CanTherm when reading scan energies from a
Gaussian log file, in case you want to see or modify the results.

The file format is a two-column table with the angles and energies.
The units are read from the first line, and can take several values.
When loading, the energies will be offset so that the first value is
zero if not already done; this way you can provide the raw values in
Hartrees if you wish.