Don't combine template reactions when loading Chemkin files.
Template reactions are usually separate transition states, and so should not
be combined into a single reaction. This is needed to handle reactions with
the same reactants and products that come from different families, or from
the same family but with different templates (e.g. different TS ring sizes).
Otherwise some Chemkin files will not parse.