Parse the HIGH keyword in Chemkin files.
Some Chemkin files put the low-pressure limit Arrhenius parameters on the
reaction line, and use the HIGH keyword to specify the high-pressure limit
parameters. This commit enables our Chemkin parser to handle that case. Note
that we need to re-parse the A factor from the reaction line to get the
units correct, as the low-P units are one order higher than the high-P
units.