Proper handling of zero-point energy in CanTherm jobs.
Previously you had to specify the ground-state energy (including ZPE) if
giving a numerical value in the species parameter files. This was less than
ideal since virtually all codes give the V=0 energy (without ZPE) as their
output. This commit changes the way the energy is loaded to better handle
the loading of the ZPE and the scaling with the frequency scale factor.