Updated data/thermoTest.py with new data from RMG-Java.
I a not sure when or why the estimates stored in this test file
stopped being the same as the estimates we get with either RMG-Py
or RMG-Java, but I replaced them with what RMG-Java *currently* gives.
I.e. if it passes these tests, we are once again the same as RMG-Java.

Most of them now pass, but notice the symmetry number changes for two of the
cyclic molecules c1ccc2ccccc2c1  and C1C=CC=C1. I think we currently
get those wrong.

Also changed the checking to check we're within 0.1 kcal/mol, 0.1 cal/mol/K,
instead of 1% or whatever it used to be.