Remove unnecessary and wild imports.
When a module imports stuff it doesn't need, 
and then another module imports * from it,
then the latter module has stuff without importing it,
and it all gets very confusing.

My aim is that everything explicitly imports just what it needs,
and it's clear where things came from and when.

PyDev is proving somewhat useful in this endeavour, although it's still
a bit confused by Cython.

The unit tests and a couple of example jobs still run as before this commit.
Hopefully nothing in the website or other projects are relying on
"lucky" wild imports.