Added verboseComments option for RMG jobs.
If False (the default), RMG-Py will now discard the (often lengthy) kinetics
comments for estimated values in favor of much shorter strings with only the
most important information. This immediately frees up ~10% of the total
memory usage of an RMG job, as well as makes the saved output files much
smaller in size and easier to read. If True the full comments are retained,
as before. You set this in the options() block of an RMG-Py input file.

I implemented this inside CoreEdgeReactionModel.getKinetics() so that the
database always returns the long comment, and we only discard that comment
in RMG jobs. I don't think assembling the long comments is much of a CPU
penalty, so I've opted for a more modular design.