Added cantherm target for building CanTherm dependencies only.
This should hopefully enable people to download RMG-Py and compile it in a
mode that works for running CanTherm. In particular, we want to avoid the
tricky dependencies (e.g. OpenBabel). You'll still need all of the
dependencies that you would for a standalone CanTherm, of course: cython,
numpy, scipy, matplotlib, and maybe one or two others.