Turned solver/liquid.pyx into a Liquid Phase Batch Reactor.
There is no pressure, and you initialize it with species
concentrations (IN SI UNITS!) not mole fractions.

the volume is calculated at the beginning so that you contain
one mole total of initial core species (like for the ideal gas reactor
where we have used mole fractions as initial mole number) but
then the volume is stored and held constant. This is then used
to calculate concentrations as the simulation proceeds.

This was based on a couple of cherry-picked commits that has been modified
by @rwest Richard West <r.west@neu.edu> to reflect recent changes