Also check for rate rules with swapped template for R_Recombination reactions.
Rate rules could be specified in either the A + B or B + A order in the
database. I'm not aware of any particular convention for this (or even if
one could be enforced), so we need to check both when looking up the
kinetics. If only one of them gives an exact match, then we will keep that
one. If both are exact or both are estimates, then we keep the one in the
order given; although I don't think this is the best way to do it.

An example is H + CH3 -> CH4, which has an exact rule for (C_methyl,H_rad)
but not for (H_rad,C_methyl). Now you will get the exact rule no matter
which order you specify the template in.