Fix SimpleReactor to use isobaric model.
The new implementation uses total moles instead of concentration for the
y(t) vector. The volume is not stored in y(t) since it can be computed
directly via the ideal gas law from the total moles, and we don't need to
compute dV/dt in this formulation.

I think the previous implementation was actually isochoric instead of
isobaric. Note that the minimal example returns the same result in either
case.