Added scoop example input file and LSF submission file.
The example is a copy of the methylformate example.
The lsf.sh script should be submitted to the LSF queuing system.
This is the system used on "Venture" at Northeastern University.

Submit with "bsub < lsf.sh".

Apparently Lava is related, so it may work for that also.
http://en.wikipedia.org/wiki/Platform_LSF

Equivalent ones should probably be written for PBS, etc.

NB. the RMG_DB_FILE environment variable is required.