Fixed bug in drawing of Network PES.
The bug occurred when trying to condense the space occupied by reactant
channels (placed to the left of the unimolecular isomers). The condensing
caused the reactant isomers to move to the right; we just needed to shift
the entire PES back to the left so that it actually appears in the rendering
window.