Can now prune network in convertFAME.py using -x flag.
You use -x to specify an energy (magnitude and units) above which any
configurations or reaction barrier heights are removed during the conversion
of a FAME input file to a CanTherm input file. This helps crop some of the
more bizarre species that RMG-Java comes up with (e.g. triradicals), leading
to a more tractable PES.