Don't use Conformers just to store E0 from the Thermo data.
Rather than create a Conformer as a side effect of getting thermo data,
we now just make sure the thermo data has an E0 attached to it
(this requires fitting a Wilhoit if, for example, you read a NASA polynomial
in from a database, and want to keep it as a NASA polynomial).

We now only create Conformers if and when needed in the generateStatMech()
method rather than the generateThermoData() method.