Pickling and unpickle database for scoop.
This is the start of a framework for sharing the database
across multiple workers with scoop.

Saving is a method of rmgpy.data.rmg.
Loading is done when needed.

The filename used to store the database pickle is set via the environment variable
RMG_DB_FILE.

Call it with something like:
 RMG_DB_FILE=$PWD/database.pkl python -m scoop -vv rmg.py path/to/input.py