Added rmgpy.thermo section to API Reference.
The organization is based somewhat on that used by scipy. The index page
contains an overview of the classes and functions defined in the subpackage,
with links to separate standalone pages for each class and function. The
toctree is hidden in favor of an alternative (table-based) means of
specifying the classes and functions.

You can still supplement the in-code docstrings with extra content, but note
that this will appear after the in-code docstring in the generated HTML.