Added table with module overview to API Reference main page.
The table shows a brief description of the main purpose of each component
(module or subpackage). You can click on the component name to view more
information about that component.

With this commit, we have completely hidden all toctrees in the API
Reference in favor of the manual, table-based navigation. This perhaps
requires a bit more effort to maintain, but I think it's a cleaner
presentation of the API overall.