Force barriers to be >=0 for all reactions that are run through PDep code.
Previously, kinetics coming from a reaction library (such as Glarborg/HighP)
were exempt from the fixBarrierHeight check. Now, if we are going to 
run them through a PDep calculation (PDep is on and the reaction is 
unimolecular) then we fix the barrier height.

Should address #77, or at least the comment
https://github.com/GreenGroup/RMG-Py/issues/77#issuecomment-14205741