Raise barriers for Library and Seed reactions to avoid PDep problem.
Still addressing issue #77.
Commit f477d79fc38f36b5014b50d0efb4ec855878b8ae added a lot of what's needed,
but the fixBarrierHeight was not being called on Seed Mechanisms or Reaction Libraries,
because these avoid the model.enlarge() method.

This commit adds that, calculating the thermo when necessary,
and also limits it to only working on Arrhenius reactions, which
have an Ea to change.

Remaining potential problem: Some high-P libraries have things like MultiArrhenius reactions.