Tweak the PDep grainCount improving error checking.
The "give up if you're getting worse" check was too stringent and it gave 
up on a job that would have succeeded had it tried one more iteration. 
This is a difficult thing to tune - I'm not sure how we should go about it.

Perhaps keep an eye on free memory and keep increasing grains until we're close 
to filling it? Anyway, this tweak helps the currently problematic system, and 
we'll just have to keep an eye out for more.