Add sensitivity option during simulation within the solver.
SimpleReactor now has keyword 'sensitivity', which can be set to
True in the input file in order to run sensitivities.  All species
sensitivities will be saved to an external xls file simulation_final.xls if
'saveConcentrationProfiles' is set to True under 'options'. Sensitivity
is run after model generation is completed on the final model.

Evaluation of the jacobian is called at every iteration when sensitivity is on,
otherwise the jacobian is not corrected often enough and causes sensitivity to
be inaccurate.  This makes running sensitivity very slow.  A faster solution
is necessary.