Allow user to choose specific species sensitivities to save.
This prevents from saving gigantic excel files of all the species.  The user
needs to set a list of the species names for the flag 'sensitivity'
for each simpleReactor object within the input file.  'saveConcentrationProfiles'
must still be set to true for excel sheet to be created.  Sensitivity results
can be found in additional sheets in 'solver/simulation_final.xls'.