Set a sensitivity threshold value to prevent writing all sensitivities to disk.
Stores the normalized sensitivities in memory and only writes them to csv if
the maximum normalized concentration sensitivity for a particular reaction
is above the user specified threshold, or a default value of 1e-3.
Previously, all reaction sensitivities were written to disk, but this creates
csv files which are far too large. Adding this additional option matches
CHEMKIN's sensitivity module as well.