Allow people without MOPAC to run RMG.
The 'check that you can find Mopac' test was being run on importing rmgpy.qm.mopac
which apparently is done whether or not QM is being used.

Now it is postponed until you are about to run a job.
I did the same for Gaussian while I was at it.