Created a DistanceData class for storing distance geometry information.
This should be usable for the group additive distance geometry estimates.
Currently it's defined in the rmgpy.data.kinetics module in the transitionstates file,
which may cause some cyclic import difficulties. Perhaps it should live elsewhere
(rmgpy.kinetics.KineticsData is an analogy)

The add(other) method should be helpful for the group ADDitive method.
Hopefully copy.copy(distancedata) will work as well (but will it need a deepcopy?)