Some implementing of getTS option to help double ended searches.
This would make all calls to __generateProductStructures() also return
the "transition state structure", a list containing [merged_reactants, merged_products]
with the atom ordering consistent (needed for double-ended searches).

This is incomplete, but along the lines of what used to be implemented.

Abandoning for now in favor of group-additive single-ended approach