Fixing math overflow error for Cv (vibrational)
Sometimes when the frequency is large it results in
exp(Beta*h*freq) to cause a math overflow. For now
one of the fix is to assume Cv = 0 when x > 500.

If exp(x) >> 1 then
Cv += x * x * exp(-x)
For values of x > 500 this should be very small number
and it is safe to assume it is 0.